Refactor : test를 위해 폴더 구조 변경
test를 위해 하나의 파일에 있던 함수들을 모두 하나의 개별 파일로 만들어줌
